% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_session.R
\name{server_session}
\alias{server_session}
\title{Session management for HIDAP-AGROFIMS}
\usage{
server_session(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Session management for HIDAP-AGROFIMS
}
\author{
Omar Benites Ivan Perez
}
