% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_session.R
\name{ui_session}
\alias{ui_session}
\title{shiny UI session management in HIDAP-AGROFIMS}
\usage{
ui_session(type = "tab", title = "User session",
  name = "open_session")
}
\arguments{
\item{type}{type of ui element; default is a tab in a shiny dashboard}

\item{title}{display title}

\item{name}{a reference name}
}
\description{
shiny UI session management in HIDAP-AGROFIMS
}
\author{
Omar Benites Ivan Perez
}
